No longer rely on the release infrastructure repo
The paths inside that repo change, but the contents of the service
definition stay largely the same. Therefore it is better to just
use a copy and maintain this copy going forward.

Kubernetes can be run with any kind of service definiton, so this is
not part of Kubernetes proper.